backend/pyproject.toml
[project]
name = "trip-splitter-backend"
version = "0.1.0"
description = "FastAPI backend for Trip Splitter expense sharing app"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "httpx",
    "anthropic",
    "pydantic-settings",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
